+++ examples/get_scan_by.rs
@@ -35,7 +35,9 @@ fn main() -> io::Result<()> {
 
     // dbg!(spectrum);
     println!("ID: {}; Index: {}; Time: {}", spectrum.id(), spectrum.index(), spectrum.start_time());
-    println!("Num data points: {}", spectrum.raw_arrays().unwrap().mzs().unwrap().len());
-
+    println!("Num data points: {}", spectrum.raw_arrays().and_then(|r| Some(r.mzs().ok()?.len())).unwrap_or_default());
+    if spectrum.signal_continuity().is_centroid() && spectrum.raw_arrays().is_none() {
+        println!("Num peaks: {}", spectrum.peaks().len());
+    }
     Ok(())
 }