+++ src/spectrum/bindata/map.rs
@@ -457,6 +457,7 @@ macro_rules! _populate_stacked_array_from {
                 bin_array.push(v)?;
             } else {
                 let mut bin_array = DataArray::from_name_and_type($array_type, $array.dtype());
+                *bin_array.unit_mut() = $array.unit();
                 bin_array.push(v)?;
                 bin.add(bin_array);
             }
@@ -629,7 +630,7 @@ impl BinaryArrayMap3D {
         for (layer, im) in self.arrays.iter().zip(self.ion_mobility_dimension.iter()) {
             sizes.clear();
             for (key, array) in layer.iter() {
-                match destination.get_mut(&key) {
+                match destination.get_mut(key) {
                     Some(sink) => {
                         sink.extend_raw(&array.data).inspect_err(|e| {
                             log::error!("Failed to extend {key:?}: {e}");
@@ -647,10 +648,8 @@ impl BinaryArrayMap3D {
             }
             if let Some(mz_size) = mz_size {
                 im_dim.extend_iter(std::iter::repeat_n(*im, mz_size))?;
-            } else {
-                if let Some((_, size)) = sizes.first() {
-                    im_dim.extend_iter(std::iter::repeat_n(*im, *size))?;
-                }
+            } else if let Some((_, size)) = sizes.first() {
+                im_dim.extend_iter(std::iter::repeat_n(*im, *size))?;
             }
         }
 
@@ -694,6 +693,7 @@ impl BinaryArrayMap3D {
             return Err(ArrayRetrievalError::NotFound(ArrayType::IonMobilityArray));
         }
         let (im_dim, im_type) = source.ion_mobility()?;
+        this.ion_mobility_unit = source.get(&im_type).unwrap().unit;
         this.ion_mobility_type = im_type;
         if im_dim.is_empty() {
             return Ok(this);
@@ -835,6 +835,9 @@ mod test {
 
         let spec = reader.get_spectrum_by_id("merged=42869 frame=9717 scanStart=1 scanEnd=705").unwrap();
         let mut arrays = spec.arrays.unwrap();
+        let units_map: HashMap<_, _> = arrays.iter().map(|(k, v)| {
+            (k.clone(), v.unit)
+        }).collect();
         let mzs = arrays.mzs()?;
         assert!(!mzs.is_sorted());
         drop(mzs);
@@ -849,12 +852,17 @@ mod test {
             .map(|(_, va)| va.mzs().unwrap().len())
             .sum();
 
+        assert_eq!(units_map[&arrays_3d.ion_mobility_type], arrays_3d.ion_mobility_unit);
+
         assert_eq!(n, stacked_n);
 
         let unstacked = arrays_3d.unstack()?;
         let unstacked_n = unstacked.mzs()?.len();
 
         assert_eq!(unstacked_n, n);
+        for (k, v) in unstacked.iter() {
+            assert_eq!(units_map[k], v.unit);
+        }
         Ok(())
     }
 }